+++ leeger/league_loader/ESPNLeagueLoader.py
@@ -88,6 +88,7 @@ class ESPNLeagueLoader(LeagueLoader):
         if validate:
             # validate new league
             leagueValidation.runAllChecks(league)
+        self._warnForUnusedOwnerNames(league)
         return league
 
     def __buildLeague(self, espnLeagues: list[ESPNLeague]) -> League: