+++ leeger/league_loader/FleaflickerLeagueLoader.py
@@ -74,6 +74,7 @@ class FleaflickerLeagueLoader(LeagueLoader):
         if validate:
             # validate new league
             leagueValidation.runAllChecks(league)
+        self._warnForUnusedOwnerNames(league)
         return league
 
     def __buildLeague(self, fleaflickerLeagues: list[dict]) -> League: